% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDDD.R
\name{getDDD}
\alias{getDDD}
\title{Download defined daily dose (DDD) information}
\usage{
getDDD(atc, comb = TRUE, progress = FALSE)
}
\arguments{
\item{atc}{Character vector of ATC codes for which to download DDD information.}

\item{comb}{Logical. Search list of combined product DDDs?}

\item{progress}{Logical. Report progress in real time?}
}
\value{
A data frame with all available DDD information for the selected ATC codes. For combined product DDDs, the information is printed to the console.
}
\description{
Download defined daily dose (DDD) information
}
\examples{
\dontrun{
# Two regular ATC codes
getDDD(atc = c("J01CA04", "J01CA05"))
# Two regular ATC codes and a combined product ATC code
getDDD(atc = c("J01CA04", "J01CA05", "J01CA20"))
# Two regular ATC codes, a fake ATC code, and a malformed code
getDDD(atc = c("J01CA04", "J01CA05", "B01AC99", "BB1AC99"))
}
}
